Added Makefile
The makefile defaults to making everything. You can also make just
the words binary, each of the tools, and each of the dictionary
files.

In the future I should probably create a separate rule to just make
the latin part.

Additionally, each of the binaries depends on all of the soruce files
according to the makefile, which probably isn't true. This should be
fixed eventually.